move shiftDecimal to common
